Create WithCoordinatesOrgUnitsSelector to reuse it
