+++ src/webapp/components/dialogs/OrganisationUnitDialog.tsx
@@ -1,8 +1,8 @@
-import React, { useEffect } from "react";
-import { OrgUnitsSelector, ConfirmationDialog } from "d2-ui-components";
-import { useAppContext, useCompositionRoot } from "../../contexts/app-context";
+import React from "react";
+import { ConfirmationDialog } from "d2-ui-components";
 import { DialogContent } from "@material-ui/core";
 import i18n from "@dhis2/d2-i18n";
+import WithCoordinatesOrgUnitsSelector from "../org-unit/WithCoordinatesOrgUnitsSelector";
 
 interface OUDialogProps {
     selectedOUs: string[];
@@ -12,22 +12,7 @@ interface OUDialogProps {
 
 const OUDialog: React.FC<OUDialogProps> = props => {
     const { selectedOUs, onCancel, onSave } = props;
-    const { api } = useAppContext();
     const [selectedOrgs, setSelectedOrgs] = React.useState<string[]>(selectedOUs);
-    const [selectablesOrgs, setSelectablesOrgs] = React.useState<string[]>([]);
-    const orgUnits = useCompositionRoot().orgUnits();
-
-    useEffect(() => {
-        orgUnits.getWithCoordinates.execute().then(orgUnits => {
-            setSelectablesOrgs(orgUnits.map(ou => ou.id));
-        });
-    }, [orgUnits.getWithCoordinates]);
-
-    const controls = {
-        filterByLevel: true,
-        filterByGroup: true,
-        selectAll: true,
-    };
 
     return (
         <React.Fragment>
@@ -41,14 +26,9 @@ const OUDialog: React.FC<OUDialogProps> = props => {
                 fullWidth={true}
             >
                 <DialogContent>
-                    <OrgUnitsSelector
-                        api={api}
-                        typeInput="checkbox"
-                        fullWidth={true}
-                        controls={controls}
+                    <WithCoordinatesOrgUnitsSelector
                         onChange={setSelectedOrgs}
                         selected={selectedOUs ? selectedOrgs : []}
-                        selectableIds={selectablesOrgs}
                     />
                 </DialogContent>
             </ConfirmationDialog>
